BUG#53421 Part of transaction not written in binlog after deadlock, replication breaks
When a "CREATE TEMPORARY TABLE SELECT * FROM" was executed the OPTION_KEEP_LOG was
not set into the thd->variables.option_bits. For that reason, if the transaction
had updated only transactional engines and was rolled back at the end (.e.g due to
a deadlock) the changes were not written to the binary log, including the creation
of the temporary table.
      
To fix the problem, we have set the OPTION_KEEP_LOG into the
thd->variables.option_bits when a "CREATE TEMPORARY TABLE
SELECT * FROM" is executed.
--BZR--
revision-id: alfranio.correia@sun.com-20100601143907-rv6ebizhuk0qyrxk
property-branch-nick: mysql-trunk-bugfixing
testament3-sha1: 25e02d71bdc1f5ace38cccc86cb90a745c3e9333