BUG#53893: RBR: nullable unique key can lead to out-of-sync slave
Post-push fix.
  
There was a valgrind issue on the loop that checks whether there
are NULL fields in the UNIQUE KEY or not. In detail, for the last 
iteration the server may read out of the key_part array boundaries,
making valgrind to output warnings.

We fix this by correcting the loop, ie, moving the part that reads
from the key_part to be inside the loop statement block. This way
the assignment is protected by the loop condition.

--BZR--
revision-id: luis.soares@sun.com-20100603234507-89huq6t3732830jg
property-branch-nick: mysql-5.1-bugteam-gca
testament3-sha1: 65fd29d8bfdc64acba234e356af8b85b9eeba326