Fix bug#27072:  RPM autostarts the server
This is the fix for 5.1, where only the behaviour on upgrade is changed:
If the server was stopped when the upgrade begins, we assume the
administrator is taking manual action, so we do not start the (new)
server at the end of the upgrade.
We still install the start/stop script, so it will be started on reboot.

--BZR--
revision-id: joerg@mysql.com-20100615202748-zh9uh5ekdnduzuaj
property-branch-nick: bug27072-5.1
property-file-info: ld7:file_id66:sp1f-mysql.spec.sh-19700101030959-man6e3acwxvf62bdqvkpcpsvdtokf3ff7:message343:In the "pre" section of the spec file, check the server status, and write
property-file-info: the result to a file.
property-file-info: In the "post" section, evaluate the status file, and start the server if
property-file-info: it was running during status analysis.
property-file-info: In 5.1, we start the server if there is no status file (which will happen
property-file-info: on first installation, when there is no data directory yet).
property-file-info: 4:path27:support-files/mysql.spec.shee
testament3-sha1: fefa6c31db0e241e85255abe6c477779eeae5544