BUG#54509: rpl_show_slave_running crashes the server sporadically
Problem: SQL and IO thread were racing for the IO_CACHE. The former to
flush it, the latter to close it. In some cases this would cause the
SQL thread to lock an invalid IO_CACHE mutex (it had been destroyed by
IO thread). This would happen when SQL thread was initializing the
master.info

Solution: We solve this by locking the log and checking if it is
hot. If it is we keep the log while seeking. Otherwise we release it
right away, because a log can get from hot to cold, but not from cold
to hot.
--BZR--
revision-id: luis.soares@sun.com-20100624180323-pzyt6xdae587zaa6
property-branch-nick: mysql-trunk-bugfixing
testament3-sha1: d6b665725477971496135b735c319dc705622b8d