BUG#54744: valgrind reports leak for mysqlbinlog
The server was not cleaning up some dbug allocated memory 
before exiting. This is not a real problem, as this memory 
would be deallocated anyway. Nonetheless, we improve the 
mysqlbinlog exit procedure, wrt to memory book-keeping, when 
no parameter is given.
      
To fix this, we deploy a call to my_end() before the
thread exits.
--BZR--
revision-id: luis.soares@sun.com-20100708224048-1rq5sfm4ssbfh0du
property-branch-nick: mysql-trunk-bugfixing
testament3-sha1: adb6bfbe0d2b57480b304287a6e2d56e620f4155