Backport of the fix for bug#25421 to 5.0.
Calculating the estimated number of records for a range scan
may take a significant time, and it was impossible for a user
to interrupt that process by killing the connection or the
query.

Fixed by checking the thread's 'killed' status in
check_quick_keys() and interrupting the calculation process if
it is set to a non-zero value.

--BZR--
revision-id: alexey.kopytov@sun.com-20100715061016-xl3hlrqfrb844bn4
property-branch-nick: mysql-5.0-bugteam
testament3-sha1: 633f5b9bbd024a352b919bb81ac6a96623ba8d39