Bug #54734 assert in Diagnostics_area::set_ok_status
This assert checks that the server does not try to send OK to the
client if there has been some error during processing. This is done
to make sure that the error is in fact sent to the client.

The problem was that view errors during processing of WHERE conditions
in UPDATE statements where not detected by the update code. It therefore
tried to send OK to the client, triggering the assert.
The bug was only noticeable in debug builds.

This patch fixes the problem by making sure that the update code
checks for errors during condition processing and acts accordingly.

--BZR--
revision-id: jon.hauglid@oracle.com-20100719090352-40qkjnfs1ikcktpv
property-branch-nick: mysql-5.1-bugteam-bug54734
testament3-sha1: f8a09da3b4c409f484b751dcbc50ff033fe29364