Bug #55565: debug assertion when ordering by expressions with user variable assignments
The assert() that is firing is checking if expressions that can't be
null return a NULL when evaluated.
MAKEDATE() function can return NULL if the second argument is 
less then or equal to 0. Thus its nullability depends not only on 
the nullability of its arguments but also on their values.
Fixed by (overoptimistically) setting MAKEDATE() to be nullable 
despite the nullability of its arguments.
Test added.
Had to update one test result to reflect the metadata change.

--BZR--
revision-id: georgi.kodinov@oracle.com-20100813130546-l7xj1x2jywsu6rru
property-branch-nick: B55565-5.1-bugteam
testament3-sha1: ab13824e97f704698f3960147785f0d820d73002