BUG#53452 Inconsistent behavior of binlog_direct_non_transactional_updates with temp table
This patch introduces two key changes in the replication's behavior.
            
Firstly, it reverts part of BUG#51894 which puts any update to temporary tables
into the trx-cache. Now, updates to temporary tables are handled according to
the type of their engines as a regular table.
            
Secondly, an unsafe mixed statement, (i.e. a statement that access transactional
table as well non-transactional or temporary table, and writes to any of them),
are written into the trx-cache in order to minimize errors in the execution when
the statement logging format is in use.
            
Such changes has a direct impact on which statements are classified as unsafe
statements and thus part of BUG#53259 is reverted.
--BZR--
revision-id: alfranio.correia@oracle.com-20100820025958-i86igkp0hf3aiw6u
property-branch-nick: mysql-5.5-bugfixing
testament3-sha1: dbb02a8778da1638d1ff71ca5abeb4eb9368ee9e