Bug#54543: update ignore with incorrect subquery leads to assertion failure: inited==INDEX
When an error occurs while sending the data in a temporary table there was no
cleanup performed. This caused a failed assertion in the case when different
access methods were used for populating the table vs. retrieving the data from
the table if IGNORE was specified and sql_safe_updates = 0. In this case
execution continues, but the handler expects to continue with the access
method used for row retrieval.

Fixed by doing the cleanup even if errors occur.
--BZR--
revision-id: martin.hansson@oracle.com-20100907075805-bdwvuzoxl1jrzlfz
property-branch-nick: 5.1bt-commit
testament3-sha1: 02c0adb739db37b2766173748151a0e867fe7705