Fix bug#56574: After installation from RPM, server is run under root, not mysql user
The problem was that in the cmake way of building
the variable "MYSQLD_USER" was not set and propagated.
In the script "mysqld_safe" its value is used as the
name of the user who should run the server process.

The fix is to explicitly set this variable to "mysql"
and propagate it in the build process.
It was analyzed and proposed by Jonathan Perkin. 
--BZR--
revision-id: joerg@mysql.com-20100907150516-jc5b33nzr4s1cu7l
property-branch-nick: clone-5.5
testament3-sha1: 1f6fd03e9e745e25d6a3f599c7cb6dea4dd33189